feat: displays the response messages for each check
