--- main.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import applyChecklist from './script';
-
-export const Popup = (props) => {
-  applyChecklist();
-  return <div>fhsdjkfhsdujkhfnkjsdhf</div>;
-};
-
-ReactDOM.render(<Popup />, document.getElementById('app'));
