+++ manifest.json
@@ -5,7 +5,7 @@
   "manifest_version": 3,
   "permissions": ["activeTab", "tabs", "scripting"],
   "action": {
-    "default_popup": "popup.html"
+    "default_popup": "/public/popup.html"
   }
 
 }
