+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  entry: ['babel-polyfill', './main.js'],
+  entry: ['babel-polyfill', './src/main.js'],
   output: {
     path: path.resolve(path.join(__dirname, 'public')),
     filename: 'bundle.js',
